requirements.txt
python-dotenv==0.19.0
requests==2.26.0
httpcache==0.1.3
ipython==7.28.0
pigeon-jupyter==0.1.0
pandas==1.3.3